amos/crawler/requirements.txt
psycopg2-binary
certifi
chardet
idna
requests
selenium
python-dotenv
coverage
mock
pytest
Pillow
googlemaps
